fix(slo): timestamp field was allowing mutliple values (#194311)
(cherry picked from commit c4e2a7256b76cc94c08f34bb6dc3faeff58777f1)

# Conflicts:
#	x-pack/plugins/observability_solution/slo/public/pages/slo_edit/components/common/group_by_field_selector.tsx
#	x-pack/plugins/observability_solution/slo/public/pages/slo_edit/components/custom_common/index_and_timestamp_field.tsx
