+++ x-pack/plugins/observability_solution/slo/public/pages/slo_edit/components/custom_kql/custom_kql_indicator_type_form.tsx
@@ -13,9 +13,9 @@ import { GroupByField } from '../common/group_by_field';
 import { useCreateDataView } from '../../../../hooks/use_create_data_view';
 import { CreateSLOForm } from '../../types';
 import { DataPreviewChart } from '../common/data_preview_chart';
-import { IndexFieldSelector } from '../common/index_field_selector';
 import { QueryBuilder } from '../common/query_builder';
 import { DATA_VIEW_FIELD, IndexSelection } from '../custom_common/index_selection';
+import { TimestampFieldSelector } from '../common/timestamp_field_selector';
 
 export function CustomKqlIndicatorTypeForm() {
   const { watch } = useFormContext<CreateSLOForm>();
@@ -35,19 +35,13 @@ export function CustomKqlIndicatorTypeForm() {
           <IndexSelection selectedDataView={dataView} />
         </EuiFlexItem>
 
-        <IndexFieldSelector
-          indexFields={timestampFields}
-          name="indicator.params.timestampField"
-          label={i18n.translate('xpack.slo.sloEdit.timestampField.label', {
-            defaultMessage: 'Timestamp field',
-          })}
-          placeholder={i18n.translate('xpack.slo.sloEdit.timestampField.placeholder', {
-            defaultMessage: 'Select a timestamp field',
-          })}
-          isLoading={!!index && isIndexFieldsLoading}
-          isDisabled={!index}
-          isRequired
-        />
+        <EuiFlexItem>
+          <TimestampFieldSelector
+            fields={timestampFields}
+            isLoading={!!index && isIndexFieldsLoading}
+            isDisabled={!index}
+          />
+        </EuiFlexItem>
       </EuiFlexGroup>
 
       <EuiFlexItem>
