+++ x-pack/plugins/observability_solution/slo/public/pages/slo_edit/components/histogram/histogram_indicator_type_form.tsx
@@ -21,10 +21,10 @@ import { useCreateDataView } from '../../../../hooks/use_create_data_view';
 import { GroupByField } from '../common/group_by_field';
 import { CreateSLOForm } from '../../types';
 import { DataPreviewChart } from '../common/data_preview_chart';
-import { IndexFieldSelector } from '../common/index_field_selector';
 import { QueryBuilder } from '../common/query_builder';
 import { DATA_VIEW_FIELD, IndexSelection } from '../custom_common/index_selection';
 import { HistogramIndicator } from './histogram_indicator';
+import { TimestampFieldSelector } from '../common/timestamp_field_selector';
 
 export function HistogramIndicatorTypeForm() {
   const { watch } = useFormContext<CreateSLOForm>();
@@ -37,7 +37,7 @@ export function HistogramIndicatorTypeForm() {
   });
 
   const histogramFields = dataView?.fields.filter((field) => field.type === 'histogram');
-  const timestampFields = dataView?.fields.filter((field) => field.type === 'date');
+  const timestampFields = dataView?.fields.filter((field) => field.type === 'date') ?? []
 
   return (
     <>
@@ -56,18 +56,10 @@ export function HistogramIndicatorTypeForm() {
             <IndexSelection />
           </EuiFlexItem>
           <EuiFlexItem>
-            <IndexFieldSelector
-              indexFields={timestampFields ?? []}
-              name="indicator.params.timestampField"
-              label={i18n.translate('xpack.slo.sloEdit.timestampField.label', {
-                defaultMessage: 'Timestamp field',
-              })}
-              placeholder={i18n.translate('xpack.slo.sloEdit.timestampField.placeholder', {
-                defaultMessage: 'Select a timestamp field',
-              })}
+            <TimestampFieldSelector
+              fields={timestampFields}
               isLoading={!!index && isIndexFieldsLoading}
               isDisabled={!index}
-              isRequired
             />
           </EuiFlexItem>
         </EuiFlexGroup>
